requirements.txt
black==23.9.0
click==8.1.7
pytest==7.4.2
pytest-cov==4.1.0
pylint==2.17.5
nbval==0.10.0

pdfplumber==0.10.2
pandas==2.1.1
oauth2client==4.1.3
gspread==5.11.3
gspread_dataframe==3.3.1
google-auth==2.23.3
google-api-python-client==2.104.0
apscheduler==3.10.4
python-dotenv==1.0.0
fastapi==0.104.0
uvicorn==0.23.2
pydantic-settings==2.0.3
python-multipart==0.0.6
jinja2==3.1.2
itsdangerous==2.1.2